requirements.txt
time
requests
RPi.GPIO
json